remove hooks for old account types
