+++ apps/wallet/src/ui/app/components/ledger/useImportLedgerAccountsMutation.ts
@@ -3,21 +3,15 @@
 
 import { useMutation, type UseMutationOptions } from '@tanstack/react-query';
 
+import { type DerivedLedgerAccount } from './useDeriveLedgerAccounts';
 import { useBackgroundClient } from '../../hooks/useBackgroundClient';
 import { type SerializedUIAccount } from '_src/background/accounts/Account';
-import { type LedgerAccountSerializedUI } from '_src/background/accounts/LedgerAccount';
-import { type SerializedLedgerAccount } from '_src/background/keyring/LedgerAccount';
 import { type Message } from '_src/shared/messaging/messages';
 
 type UseImportLedgerAccountsMutationOptions = Pick<
-	UseMutationOptions<
-		Message | SerializedUIAccount[],
-		unknown,
-		SerializedLedgerAccount[] | LedgerAccountSerializedUI[],
-		unknown
-	>,
+	UseMutationOptions<Message | SerializedUIAccount[], unknown, DerivedLedgerAccount[], unknown>,
 	'onSuccess' | 'onError'
-> & { password?: string };
+> & { password: string };
 
 export function useImportLedgerAccountsMutation({
 	onSuccess,
@@ -26,20 +20,16 @@ export function useImportLedgerAccountsMutation({
 }: UseImportLedgerAccountsMutationOptions) {
 	const backgroundClient = useBackgroundClient();
 	return useMutation({
-		mutationFn: async (ledgerAccounts: SerializedLedgerAccount[]) => {
-			if (password) {
-				return backgroundClient.createAccounts({
-					type: 'ledger',
-					accounts: ledgerAccounts.map(({ address, derivationPath, publicKey }) => ({
-						address,
-						derivationPath,
-						publicKey: publicKey!,
-					})),
-					password,
-				});
-			}
-			return backgroundClient.importLedgerAccounts(ledgerAccounts as SerializedLedgerAccount[]);
-		},
+		mutationFn: async (ledgerAccounts: DerivedLedgerAccount[]) =>
+			backgroundClient.createAccounts({
+				type: 'ledger',
+				accounts: ledgerAccounts.map(({ address, derivationPath, publicKey }) => ({
+					address,
+					derivationPath,
+					publicKey: publicKey!,
+				})),
+				password,
+			}),
 		onSuccess,
 		onError,
 	});
