+++ apps/wallet/src/ui/app/helpers/query-client-keys.ts
