+++ apps/wallet/src/ui/app/hooks/useAccounts.ts
@@ -1,20 +1,19 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { useMemo } from 'react';
+import { useQuery } from '@tanstack/react-query';
+import { useBackgroundClient } from './useBackgroundClient';
+import { accountsQueryKey } from '../helpers/query-client-keys';
+import { type SerializedUIAccount } from '_src/background/accounts/Account';
 
-import useAppSelector from './useAppSelector';
-import { accountsAdapterSelectors } from '../redux/slices/account';
-
-/**
- * @deprecated use accounts-v2/useAccounts
- */
-export function useAccounts(addressesFilters?: string[]) {
-	const accounts = useAppSelector(accountsAdapterSelectors.selectAll);
-	return useMemo(() => {
-		if (!addressesFilters?.length) {
-			return accounts;
-		}
-		return accounts.filter((anAccount) => addressesFilters.includes(anAccount.address));
-	}, [accounts, addressesFilters]);
+export function useAccounts() {
+	const backgroundClient = useBackgroundClient();
+	return useQuery({
+		queryKey: accountsQueryKey,
+		queryFn: () => backgroundClient.getStoredEntities<SerializedUIAccount>('accounts'),
+		cacheTime: 30 * 1000,
+		staleTime: 15 * 1000,
+		refetchInterval: 30 * 1000,
+		meta: { skipPersistedCache: true },
+	});
 }
