+++ apps/wallet/src/ui/app/hooks/useActiveAccount.ts
@@ -1,12 +1,19 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import useAppSelector from './useAppSelector';
-import { activeAccountSelector } from '../redux/slices/account';
+import { useMemo } from 'react';
+import { useAccounts } from './useAccounts';
 
-/**
- * @deprecated use accounts-v2/useActiveAccount
- */
 export function useActiveAccount() {
-	return useAppSelector(activeAccountSelector);
+	const { data: allAccounts } = useAccounts();
+	return useMemo(() => {
+		if (!allAccounts) {
+			return null;
+		}
+		const selected = allAccounts.find(({ selected }) => selected);
+		if (selected) {
+			return selected;
+		}
+		return allAccounts[0] || null;
+	}, [allAccounts]);
 }
