+++ apps/wallet/src/ui/app/hooks/useActiveAddress.ts
@@ -1,12 +1,8 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import useAppSelector from './useAppSelector';
-import { activeAddressSelector } from '../redux/slices/account';
+import { useActiveAccount } from './useActiveAccount';
 
-/**
- * @deprecated use accounts-v2/useActiveAccount
- */
 export function useActiveAddress() {
-	return useAppSelector(activeAddressSelector);
+	return useActiveAccount()?.address || null;
 }
