+++ apps/wallet/src/ui/app/hooks/useSigner.ts
@@ -1,44 +1,37 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { useAccounts } from './useAccounts';
-import { useActiveAccount } from './useActiveAccount';
+import { useRpcClient } from '@mysten/core';
 import useAppSelector from './useAppSelector';
+import { useBackgroundClient } from './useBackgroundClient';
 import { useQredoAPI } from './useQredoAPI';
 import { LedgerSigner } from '../LedgerSigner';
 import { QredoSigner } from '../QredoSigner';
 import { type WalletSigner } from '../WalletSigner';
 import { useSuiLedgerClient } from '../components/ledger/SuiLedgerClientProvider';
-import { thunkExtras } from '_redux/store/thunk-extras';
-import { AccountType } from '_src/background/keyring/Account';
+import { thunkExtras } from '../redux/store/thunk-extras';
+import { type SerializedUIAccount } from '_src/background/accounts/Account';
+import { isLedgerAccountSerializedUI } from '_src/background/accounts/LedgerAccount';
+import { isQredoAccountSerializedUI } from '_src/background/accounts/QredoAccount';
 
-/**
- * @deprecated use accounts-v2/useSigner
- */
-export function useSigner(address?: string): WalletSigner | null {
-	const activeAccount = useActiveAccount();
-	const existingAccounts = useAccounts();
-	const signerAccount = address
-		? existingAccounts.find((account) => account.address === address)
-		: activeAccount;
-
+export function useSigner(account: SerializedUIAccount | null): WalletSigner | null {
 	const { connectToLedger } = useSuiLedgerClient();
-	const { api, background } = thunkExtras;
+	const api = useRpcClient();
+	const background = useBackgroundClient();
 	const [qredoAPI] = useQredoAPI(
-		signerAccount?.type === AccountType.QREDO ? signerAccount.qredoConnectionID : undefined,
+		account && !account?.isLocked && isQredoAccountSerializedUI(account)
+			? account.sourceID
+			: undefined,
 	);
 	const networkName = useAppSelector(({ app: { apiEnv } }) => apiEnv);
-	if (!signerAccount) {
-		throw new Error("Can't find account for the signer address");
+	if (!account) {
+		return null;
 	}
-
-	if (signerAccount.type === AccountType.LEDGER) {
-		return new LedgerSigner(connectToLedger, signerAccount.derivationPath, api.instance.fullNode);
+	if (isLedgerAccountSerializedUI(account)) {
+		return new LedgerSigner(connectToLedger, account.derivationPath, api);
 	}
-	if (signerAccount.type === AccountType.QREDO) {
-		return qredoAPI
-			? new QredoSigner(api.instance.fullNode, signerAccount, qredoAPI, networkName)
-			: null;
+	if (isQredoAccountSerializedUI(account)) {
+		return qredoAPI ? new QredoSigner(api, account, qredoAPI, networkName) : null;
 	}
-	return api.getSignerInstance(signerAccount, background);
+	return thunkExtras.api.getSignerInstance(account, background);
 }
