+++ apps/wallet/src/ui/app/hooks/useTransactionDryRun.ts
@@ -3,14 +3,15 @@
 
 import { type TransactionBlock } from '@mysten/sui.js/transactions';
 import { useQuery } from '@tanstack/react-query';
-
-import { useSigner } from '_hooks';
+import { useAccountByAddress } from './useAccountByAddress';
+import { useSigner } from './useSigner';
 
 export function useTransactionDryRun(
 	sender: string | undefined,
 	transactionBlock: TransactionBlock,
 ) {
-	const signer = useSigner(sender);
+	const { data: account } = useAccountByAddress(sender);
+	const signer = useSigner(account || null);
 	const response = useQuery({
 		// eslint-disable-next-line @tanstack/query/exhaustive-deps
 		queryKey: ['dryRunTransaction', transactionBlock.serialize()],
