+++ apps/wallet/src/ui/app/pages/accounts-dev/index.tsx
@@ -18,11 +18,11 @@ import LoadingIndicator from '../../components/loading/LoadingIndicator';
 import Logo from '../../components/logo';
 import NetworkSelector from '../../components/network-selector';
 import { useAppSelector } from '../../hooks';
-import { useAccountSources } from '../../hooks/accounts-v2/useAccountSources';
-import { useAccounts } from '../../hooks/accounts-v2/useAccounts';
-import { useSigner } from '../../hooks/accounts-v2/useSigner';
+import { useAccountSources } from '../../hooks/useAccountSources';
+import { useAccounts } from '../../hooks/useAccounts';
 import { useBackgroundClient } from '../../hooks/useBackgroundClient';
 import { useQredoTransaction } from '../../hooks/useQredoTransaction';
+import { useSigner } from '../../hooks/useSigner';
 import { ImportLedgerAccountsPage } from '../../pages/accounts/ImportLedgerAccountsPage';
 import { Button } from '../../shared/ButtonUI';
 import { ModalDialog } from '../../shared/ModalDialog';
