+++ apps/wallet/src/ui/app/pages/approval-request/transaction-request/index.tsx
@@ -10,12 +10,14 @@ import { GasFees } from './GasFees';
 import { TransactionDetails } from './TransactionDetails';
 import { ConfirmationModal } from '../../../shared/ConfirmationModal';
 import { UserApproveContainer } from '_components/user-approve-container';
-import { useAppDispatch, useSigner, useTransactionData, useTransactionDryRun } from '_hooks';
+import { useAppDispatch, useTransactionData, useTransactionDryRun } from '_hooks';
 import { type TransactionApprovalRequest } from '_payloads/transactions/ApprovalRequest';
 import { respondToTransactionRequest } from '_redux/slices/transaction-requests';
 import { ampli } from '_src/shared/analytics/ampli';
+import { useAccountByAddress } from '_src/ui/app/hooks/useAccountByAddress';
 import { useQredoTransaction } from '_src/ui/app/hooks/useQredoTransaction';
 import { useRecognizedPackages } from '_src/ui/app/hooks/useRecognizedPackages';
+import { useSigner } from '_src/ui/app/hooks/useSigner';
 import { PageMainLayoutTitle } from '_src/ui/app/shared/page-main-layout/PageMainLayoutTitle';
 import { TransactionSummary } from '_src/ui/app/shared/transaction-summary';
 
@@ -31,7 +33,8 @@ const appOriginsToExcludeFromAnalytics = ['https://sui8192.ethoswallet.xyz'];
 
 export function TransactionRequest({ txRequest }: TransactionRequestProps) {
 	const addressForTransaction = txRequest.tx.account;
-	const signer = useSigner(addressForTransaction);
+	const { data: accountForTransaction } = useAccountByAddress(addressForTransaction);
+	const signer = useSigner(accountForTransaction);
 	const dispatch = useAppDispatch();
 	const transaction = useMemo(() => {
 		const tx = TransactionBlock.from(txRequest.tx.data);
