+++ apps/wallet/src/ui/app/pages/home/nft-transfer/TransferNFTForm.tsx
@@ -11,16 +11,17 @@ import { useNavigate } from 'react-router-dom';
 
 import { useTransferKioskItem } from './useTransferKioskItem';
 import { createValidationSchema } from './validation';
-import { useActiveAddress } from '_app/hooks/useActiveAddress';
 import { Button } from '_app/shared/ButtonUI';
 import BottomMenuLayout, { Content, Menu } from '_app/shared/bottom-menu-layout';
 import { Text } from '_app/shared/text';
 import { AddressInput } from '_components/address-input';
-import { useSigner } from '_hooks';
 import { ampli } from '_src/shared/analytics/ampli';
 import { QredoActionIgnoredByUser } from '_src/ui/app/QredoSigner';
 import { getSignerOperationErrorMessage } from '_src/ui/app/helpers/errorMessages';
+import { useActiveAddress } from '_src/ui/app/hooks';
+import { useActiveAccount } from '_src/ui/app/hooks/useActiveAccount';
 import { useQredoTransaction } from '_src/ui/app/hooks/useQredoTransaction';
+import { useSigner } from '_src/ui/app/hooks/useSigner';
 
 export function TransferNFTForm({
 	objectId,
@@ -33,7 +34,8 @@ export function TransferNFTForm({
 	const rpc = useRpcClient();
 	const suiNSEnabled = useSuiNSEnabled();
 	const validationSchema = createValidationSchema(rpc, suiNSEnabled, activeAddress || '', objectId);
-	const signer = useSigner();
+	const activeAccount = useActiveAccount();
+	const signer = useSigner(activeAccount);
 	const queryClient = useQueryClient();
 	const navigate = useNavigate();
 	const { clientIdentifier, notificationModal } = useQredoTransaction();
