+++ apps/wallet/src/ui/app/pages/home/transactions/QredoPendingTransactions.tsx
@@ -5,11 +5,10 @@ import { QredoTransaction } from './QredoTransaction';
 import { ErrorBoundary } from '_components/error-boundary';
 import Loading from '_components/loading';
 import { NoActivityCard } from '_components/transactions-card/NoActivityCard';
-import { AccountType } from '_src/background/keyring/Account';
+import { isQredoAccountSerializedUI } from '_src/background/accounts/QredoAccount';
 import { type TransactionStatus } from '_src/shared/qredo-api';
 import Alert from '_src/ui/app/components/alert';
 import { useActiveAccount } from '_src/ui/app/hooks/useActiveAccount';
-import { useActiveAddress } from '_src/ui/app/hooks/useActiveAddress';
 import { useGetQredoTransactions } from '_src/ui/app/hooks/useGetQredoTransactions';
 
 const PENDING_QREDO_TRANSACTION_STATUSES: TransactionStatus[] = [
@@ -23,10 +22,10 @@ const PENDING_QREDO_TRANSACTION_STATUSES: TransactionStatus[] = [
 ];
 
 export function QredoPendingTransactions() {
-	const activeAddress = useActiveAddress();
 	const activeAccount = useActiveAccount();
-	const isQredoAccount = activeAccount?.type === AccountType.QREDO;
-	const qredoID = isQredoAccount ? activeAccount.qredoConnectionID : undefined;
+	const activeAddress = activeAccount?.address;
+	const isQredoAccount = !!(activeAccount && isQredoAccountSerializedUI(activeAccount));
+	const qredoID = isQredoAccount ? activeAccount.sourceID : undefined;
 	const {
 		data: qredoTransactions,
 		isLoading,
